[Fix] Continue fixes on type convertion (client)

+++ client/base/Api_protocol_message.cpp
@@ -2558,7 +2558,7 @@ bool Api_protocol::parseMessage(const uint8_t &packetCode,const QByteArray &data
         parseError(QStringLiteral("Procotol wrong or corrupted"),QStringLiteral("remaining data: parseMessage(%1,%2 %3) line %4")
                       .arg(packetCode)
                       .arg(QString(data.mid(0,static_cast<int>(in.device()->pos())).toHex()))
-                      .arg(QString(data.mid(static_cast<int>(in.device()->pos()),(in.device()->size()-in.device()->pos())).toHex()))
+                      .arg(QString(data.mid(static_cast<int>(in.device()->pos()),static_cast<int>(in.device()->size()-in.device()->pos())).toHex()))
                       .arg(QStringLiteral("%1:%2").arg(__FILE__).arg(__LINE__))
                       );
         return false;
