+++ client/base/render/MapVisualiserOrder.cpp
@@ -107,10 +107,10 @@ void MapVisualiserOrder::layerChangeLevelAndTagsChange(MapVisualiserOrder::Map_f
                                             uint8_t frames=framesString.toUShort();
                                             if(ms>0 && frames>1)
                                             {
-                                                if(!tempMapObject->doors.contains(QPair<uint8_t,uint8_t>(x,y)))
+                                                if(!tempMapObject->doors.contains(QPair<uint8_t,uint8_t>(static_cast<uint8_t>(x),static_cast<uint8_t>(y))))
                                                 {
                                                     MapDoor *door=new MapDoor(objects.at(index2),frames,ms);
-                                                    tempMapObject->doors[QPair<uint8_t,uint8_t>(x,y)]=door;
+                                                    tempMapObject->doors[QPair<uint8_t,uint8_t>(static_cast<uint8_t>(x),static_cast<uint8_t>(y))]=door;
                                                 }
                                             }
                                             else
